Completion of Task 4
The method checkProcessFinished() has been implemented and all associated tests are passing for the task. The function first checks if the cpu is idle or the current process is still running, returning with no changes if so. Otherwise, the endTime of the process will be updated, the process itself will be set to done, and the cpu will be returned to an idle state.